BUG: Fix overflow error.
The main change is in the function _scale_to_sampwidth: vmin and vmax are
converted to floating point before using them to rescale the data.  An example
of the error that could occur before this change is if data had type np.int8,
and the minimum and maximum values in the array were -128 and 127.  Then the
difference vmax - vmin would be computed using np.int8, resulting in the
value -1 instead of 255.  Since the end result of the calculation is floating
point, we might as well convert vmin and vmax to floating point right away,
and avoid the problem of integer overflow.

Tests are added, and the test code is refactored a bit to remove some
repeated code.

Closes gh-5.